# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bunkarr API",
    "version": "1",
    "description": "Bunkarr REST API. Authenticate with the API key from Settings > General in the X-Api-Key header (or the apikey query parameter), or with a login session cookie. Errors are {\"message\": \"...\"}: 400 invalid input, 401 not authenticated, 403 refused (a cross-site browser request, an action that needs a login session, a directory Bunkarr cannot read), 404 unknown id, 409 conflict with the current state (including safety refusals that need a confirmation or a mounted share), 502 Plex failed, 503 unavailable (the backup routes of a server without backup services, or a request stopped as the server shuts down), 500 anything else. Secrets (Plex tokens, Apprise URLs) are write-only. Job-starting endpoints answer 202 with the queued Job. Sizes are bytes, times RFC 3339.",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "apiKeyHeader": []
    },
    {
      "apiKeyQuery": []
    },
    {
      "session": []
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bunkarr_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "enabled",
          "apiUrl",
          "configKey",
          "hasUrls",
          "onFailure",
          "onWarning",
          "onSuccess"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "apprise"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "apiUrl": {
            "type": "string"
          },
          "configKey": {
            "type": "string",
            "description": "Apprise stateful mode; \"\" means stateless with the stored URLs"
          },
          "hasUrls": {
            "type": "boolean",
            "description": "The Apprise URLs themselves are never returned."
          },
          "onFailure": {
            "type": "boolean"
          },
          "onWarning": {
            "type": "boolean"
          },
          "onSuccess": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationInput": {
        "type": "object",
        "required": [
          "name",
          "apiUrl"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "kind": {
            "type": "string",
            "enum": [
              "apprise"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "apiUrl": {
            "type": "string"
          },
          "configKey": {
            "type": "string"
          },
          "urls": {
            "type": "string",
            "description": "Write-only Apprise URLs (comma or space separated); empty on update keeps the stored ones."
          },
          "onFailure": {
            "type": "boolean"
          },
          "onWarning": {
            "type": "boolean"
          },
          "onSuccess": {
            "type": "boolean"
          }
        }
      },
      "NotificationTest": {
        "type": "object",
        "description": "A notification form (apiUrl and urls or configKey required) or just the id of a saved notification.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "A saved notification: empty urls use its stored URLs; an id without apiUrl tests it as saved."
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "apprise"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "apiUrl": {
            "type": "string"
          },
          "configKey": {
            "type": "string"
          },
          "urls": {
            "type": "string"
          },
          "onFailure": {
            "type": "boolean"
          },
          "onWarning": {
            "type": "boolean"
          },
          "onSuccess": {
            "type": "boolean"
          }
        }
      },
      "TestResult": {
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal error (message redacted of secrets)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "No backup services are configured on this server, or the request was stopped (the server is shutting down)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    }
  },
  "paths": {
...
    "/notifications": {
      "get": {
        "operationId": "listNotifications",
        "summary": "List notifications",
        "responses": {
          "200": {
            "description": "Notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "post": {
        "operationId": "createNotification",
        "summary": "Add an Apprise notification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/test": {
      "post": {
        "operationId": "testNotification",
        "summary": "Send a test message",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationTest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (a failed delivery is ok: false)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "put": {
        "operationId": "updateNotification",
        "summary": "Update a notification; empty urls keep the stored ones",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Notifications"
        ],
        "description": "In stateless mode, changing apiUrl requires the Apprise URLs again (urls): stored URLs are only sent to the API they were saved with; otherwise 400."
      },
      "delete": {
        "operationId": "deleteNotification",
        "summary": "Delete a notification",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Done"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    }
  }
}